backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Environment
python-dotenv==1.0.1

# HTTP Client (for testing)
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Code Quality (optional)
black==24.10.0
flake8==7.1.1
mypy==1.14.1